% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.patch-package.R
\docType{package}
\name{arrow.patch-package}
\alias{arrow.patch}
\alias{arrow.patch-package}
\title{arrow.patch: Patch Arrow Functions for Serialising SF Columns}
\description{
Patch arrow functions for serialising data.frames that may contain sf columns.
}
\author{
\strong{Maintainer}: Anthony North \email{anthony.north@qfes.qld.gov.au}

Other contributors:
\itemize{
  \item Miles McBain \email{miles.mcbain@qfes.qld.gov.au} (\href{https://orcid.org/0000-0003-2865-2548}{ORCID}) [contributor]
}

}
\keyword{internal}
