% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.R
\name{decode_wkb}
\alias{decode_wkb}
\title{Decode WKB}
\usage{
decode_wkb(data)
}
\arguments{
\item{data}{<\code{any}> The input object}
}
\description{
Decodes well-known binary (WKB) blob columns as simple features columns
(\code{sfc}). This function is useful for decoding wkb columns that were
previously encoded with \code{\link[=encode_wkb]{encode_wkb()}}.
}
\details{
Well-known binary columns are identified by their \code{wkb} attribute.
}
\author{
Anthony North
}
