% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec.R
\name{encode_wkb}
\alias{encode_wkb}
\title{Encode WKB}
\usage{
encode_wkb(data)
}
\arguments{
\item{data}{<\code{any}> The input object}
}
\description{
Encodes simple features columns (\code{sfc}) as well-known binary (WKB) blob
columns and stores crs as attributes on these vectors. This allows for
serialisation to parquet, ipc_stream and feather.
}
\author{
Anthony North
}
