% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{patch_parquet}
\alias{patch_parquet}
\title{Patch parquet}
\usage{
patch_parquet()
}
\description{
Replaces \code{\link[arrow:read_parquet]{arrow::read_parquet()}} and \code{\link[arrow:write_parquet]{arrow::write_parquet()}}
\code{\link[=read_parquet]{read_parquet()}} and \code{\link[=write_parquet]{write_parquet()}}
}
\author{
Anthony North
}
